requirements.txt
requests>=2.0
pyaml>=3.0
lxml>=3.8.0
